+++ tests/test_propagation.py
@@ -23,7 +23,7 @@ from opentelemetry.propagate import get_global_textmap
 from opentelemetry.propagators.composite import CompositePropagator
 from opentelemetry.trace.propagation.tracecontext import TraceContextTextMapPropagator
 
-from splunk_otel.options import Options
+from splunk_otel.options import _Options
 from splunk_otel.propagators import ServerTimingResponsePropagator
 from splunk_otel.tracing import _configure_tracing
 
@@ -31,7 +31,7 @@ from splunk_otel.tracing import _configure_tracing
 class TestPropagator(TestCase):
     def test_sets_tracecontext_and_baggage_are_default_propagator(self):
         reload(propagate)
-        _configure_tracing(Options())
+        _configure_tracing(_Options())
         propagator = get_global_textmap()
         self.assertIsInstance(propagator, CompositePropagator)
         propagators = propagator._propagators  # pylint: disable=protected-access
@@ -45,7 +45,7 @@ class TestPropagator(TestCase):
     )
     def test_set_custom_propagator(self):
         reload(propagate)
-        _configure_tracing(Options())
+        _configure_tracing(_Options())
         propagator = get_global_textmap()
         self.assertIsInstance(propagator, CompositePropagator)
         propagators = propagator._propagators  # pylint: disable=protected-access
@@ -53,12 +53,12 @@ class TestPropagator(TestCase):
         self.assertIsInstance(propagators[0], W3CBaggagePropagator)
 
     def test_server_timing_is_default_response_propagator(self):
-        _configure_tracing(Options())
+        _configure_tracing(_Options())
         propagtor = get_global_response_propagator()
         self.assertIsInstance(propagtor, ServerTimingResponsePropagator)
 
     def test_server_timing_is_global_response_propagator_disabled_code(self):
-        _configure_tracing(Options(trace_response_header_enabled=False))
+        _configure_tracing(_Options(trace_response_header_enabled=False))
         self.assertIsNone(get_global_response_propagator())
 
     @mock.patch.dict(
@@ -66,7 +66,7 @@ class TestPropagator(TestCase):
         {"SPLUNK_TRACE_RESPONSE_HEADER_ENABLED": "false"},
     )
     def test_server_timing_is_global_response_propagator_disabled_env(self):
-        _configure_tracing(Options())
+        _configure_tracing(_Options())
         self.assertIsNone(get_global_response_propagator())